RA-1258 renamed search validation test folder to match agreed structure. added tests for searching by framework code.
